+++ src/mavlink_interface.cpp
@@ -225,7 +225,7 @@ void MavlinkInterface::Load() {
   }
 }
 
-void MavlinkInterface::SendSensorMessages(int time_usec) {
+void MavlinkInterface::SendSensorMessages(uint64_t time_usec) {
   mavlink_hil_sensor_t sensor_msg;
 
   sensor_msg.time_usec = time_usec;
